updates files to PySide6, added functionality for reset_database and train_face
