+++ .idea/workspace.xml
@@ -7,11 +7,15 @@
     <list default="true" id="38ebf632-8fcc-432a-b960-3dfb24f85e9e" name="Changes" comment="">
       <change beforePath="$PROJECT_DIR$/.idea/workspace.xml" beforeDir="false" afterPath="$PROJECT_DIR$/.idea/workspace.xml" afterDir="false" />
       <change beforePath="$PROJECT_DIR$/logic/facial_tracking/dialogs/add_face.py" beforeDir="false" afterPath="$PROJECT_DIR$/logic/facial_tracking/dialogs/add_face.py" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/logic/facial_tracking/dialogs/remove_face.py" beforeDir="false" afterPath="$PROJECT_DIR$/logic/facial_tracking/dialogs/remove_face.py" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/logic/facial_tracking/dialogs/reset_database.py" beforeDir="false" afterPath="$PROJECT_DIR$/logic/facial_tracking/dialogs/reset_database.py" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/logic/facial_tracking/dialogs/train_face.py" beforeDir="false" afterPath="$PROJECT_DIR$/logic/facial_tracking/dialogs/train_face.py" afterDir="false" />
       <change beforePath="$PROJECT_DIR$/logic/facial_tracking/testing_image_processor.py" beforeDir="false" afterPath="$PROJECT_DIR$/logic/facial_tracking/testing_image_processor.py" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/shared/message_prompts.py" beforeDir="false" afterPath="$PROJECT_DIR$/shared/message_prompts.py" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/shared/watch_trainer_directory.py" beforeDir="false" afterPath="$PROJECT_DIR$/shared/watch_trainer_directory.py" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/views/functions/assign_network_ptz_ui.py" beforeDir="false" afterPath="$PROJECT_DIR$/views/functions/assign_network_ptz_ui.py" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/views/functions/assign_visca_ptz_ui.py" beforeDir="false" afterPath="$PROJECT_DIR$/views/functions/assign_visca_ptz_ui.py" afterDir="false" />
       <change beforePath="$PROJECT_DIR$/views/functions/show_dialogs_ui.py" beforeDir="false" afterPath="$PROJECT_DIR$/views/functions/show_dialogs_ui.py" afterDir="false" />
       <change beforePath="$PROJECT_DIR$/views/homepage/main_window.py" beforeDir="false" afterPath="$PROJECT_DIR$/views/homepage/main_window.py" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/views/widgets/camera_widget.py" beforeDir="false" afterPath="$PROJECT_DIR$/views/widgets/camera_widget.py" afterDir="false" />
     </list>
     <option name="SHOW_DIALOG" value="false" />
     <option name="HIGHLIGHT_CONFLICTS" value="true" />
