+++ logic/facial_tracking/dialogs/add_face.py
@@ -1,6 +1,6 @@
 import os
-from PyQt6 import QtCore, QtWidgets
-from PyQt6.QtWidgets import QDialog
+from PySide6 import QtCore, QtWidgets
+from PySide6.QtWidgets import QDialog
 
 from shared import constants
 from shared.message_prompts import show_critical_messagebox, show_info_messagebox
