+++ shared/message_prompts.py
@@ -1,4 +1,4 @@
-from PyQt6.QtWidgets import QMessageBox
+from PySide6.QtWidgets import QMessageBox
 
 
 def show_critical_messagebox(window_title, critical_message):
