+++ views/functions/assign_network_ptz_ui.py
@@ -1,7 +1,7 @@
 import re
 from sensecam_control import onvif_control
-from PyQt6 import QtCore, QtWidgets
-from PyQt6.QtWidgets import QDialog
+from PySide6 import QtCore, QtWidgets
+from PySide6.QtWidgets import QDialog
 
 from shared.message_prompts import show_critical_messagebox
