+++ views/functions/show_dialogs_ui.py
@@ -3,6 +3,7 @@ import shutil
 
 from logic.facial_tracking.dialogs.add_face import AddFaceDlg
 from logic.facial_tracking.dialogs.remove_face import RemoveFaceDlg
+from logic.facial_tracking.dialogs.reset_database import ResetDatabaseDlg
 from logic.facial_tracking.dialogs.train_face import TrainerDlg
 from shared.message_prompts import show_info_messagebox
 import shared.constants as constants
@@ -41,4 +42,21 @@ class ShowDialog:
                 if os.path.exists(constants.ENCODINGS_PATH):
                     os.remove(constants.ENCODINGS_PATH)
             elif current_len is not len(os.listdir(constants.IMAGE_PATH)):
-                TrainerDlg().show()
+                dlg = TrainerDlg()
+                dlg.exec()
+
+    @staticmethod
+    def retrain_face():
+        if not os.path.isdir(constants.IMAGE_PATH) or not os.listdir(constants.IMAGE_PATH):
+            show_info_messagebox("No Faces to train.")
+        else:
+            dlg = TrainerDlg()
+            dlg.exec()
+
+    @staticmethod
+    def reset_database(update_face_selection):
+        """Launch the Remove Face dialog based on the currently selected camera."""
+        print("Opening Face Dialog")
+        dlg = ResetDatabaseDlg()
+        dlg.closeEvent = update_face_selection
+        dlg.exec()
